[jax2tf] Moved text from the documentation into the limitations for custom checks
* cleaned up the handling of tolerances and custom asserts.
* Removed the harnesss field from the limitations.
* Moved the definitions of the Jax2TfLimitation into its own
  file, so it can be reused.
