+++ jax/experimental/jax2tf/tests/primitives_test.py
@@ -27,7 +27,7 @@ that we get the same result as in JAX
 Some harnesses need specific tolerances, or even custom equality assertions.
 Also, for some harnesses we need to specify some data types that result
 in Tensorflow errors (for some devices and compilation modes). These limitations
-are captured as Jax2TfLimitation objects.
+are captured as jax2tf_limitations.Jax2TfLimitation objects.
 
 From the limitations objects, we generate a
 [report](https://github.com/google/jax/blob/master/jax/experimental/jax2tf/g3doc/primitives_with_limited_support.md).
@@ -41,20 +41,19 @@ turn these warnings into errors, you'd have to uncomment an assertion
 in `tf_test_util.ConvertAndCompare`.
 
 IMPORTANT: If you need to customize the testing of a particular primitive
-conversion, you must create a class method in Jax2TfLimitation,
+conversion, you must create a class method in jax2tf_limitations.jax2tf_limitations,
 with the same name as the harness.group_name (typically the same as the
-primitive name). That class method should return the list of Jax2tfLimitation
+primitive name). That class method should return the list of Jax2TfLimitation
 objects for the harness.
-See `Jax2TfLimitation.limitations_for_harness`. If a group name does
+See `jax2tf_limitations.limitations_for_harness`. If a group name does
 not need limitations, then it must be listed in the
-`Jax2TfLimitation.harness_groups_no_limitations`.
+`jax2tf_limitations.harness_groups_no_limitations`.
 
 """
 
 import datetime
-import itertools
 import os
-from typing import Any, Callable, Dict, Optional, Sequence
+from typing import Any, Dict, Tuple
 import unittest
 
 from absl.testing import absltest
@@ -67,7 +66,6 @@ from jax import numpy as jnp
 from jax import test_util as jtu
 from jax.config import config
 from jax.experimental import jax2tf
-from jax.experimental.jax2tf.tests import tf_test_util
 from jax.interpreters import xla
 
 import numpy as np
@@ -76,7 +74,11 @@ config.parse_flags_with_absl()
 FLAGS = config.FLAGS
 
 # Import after parsing flags
+from jax.experimental.jax2tf.tests import tf_test_util
+from jax.experimental.jax2tf.tests.jax2tf_limitations import Jax2TfLimitation
 from jax.experimental.jax2tf.tests import primitive_harness
+
+DType = Any
 
 REDUCE = (
     jnp.all,
@@ -92,24 +94,21 @@ class JaxPrimitiveTest(tf_test_util.JaxToTfTestCase):
 
   # This test runs for all primitive harnesses. For each primitive "xxx" the
   # test will be called "test_prim_xxx_..." and the custom parameters for
-  # the test are defined in the class method "Jax2TfLimitation.xxx".
+  # the test are defined in the class method "jax2tf_limitations.Jax2TfLimitation.xxx".
   # See more details in the comment at top of file and in Jax2TfLimitation class.
   @primitive_harness.parameterized(
-      primitive_harness.all_harnesses, include_jax_unimpl=False)
+      primitive_harness.all_harnesses, include_jax_unimpl=False,
+      )
   @jtu.ignore_warning(
       category=UserWarning, message="Using reduced precision for gradient.*")
   def test_prim(self, harness: primitive_harness.Harness):
     limitations = Jax2TfLimitation.limitations_for_harness(harness)
+    device = jtu.device_under_test()
+    limitations = tuple(filter(lambda l: l.filter(device=device,
+                                                  dtype=harness.dtype), limitations))
     func_jax = harness.dyn_fun
     args = harness.dyn_args_maker(self.rng())
-
-    def filter_limitations(device_under_test: str,
-                           mode: str) -> Sequence["Jax2TfLimitation"]:
-      return [
-          l for l in limitations
-          if (mode in l.modes and l.filter(device_under_test))]
-
-    self.ConvertAndCompare(func_jax, *args, limitations=filter_limitations)
+    self.ConvertAndCompare(func_jax, *args, limitations=limitations)
 
   def test_primitive_coverage(self):
     """Fail if there are JAX primitives that are not implemented."""
@@ -149,25 +148,24 @@ class JaxPrimitiveTest(tf_test_util.JaxToTfTestCase):
     ]
     print(f"Found {len(harnesses)} test harnesses that work in JAX")
 
-    def unique_hash(l: Jax2TfLimitation):
-      return hash(
-          (l.harness.group_name, l.description, l.devices, l.dtypes, l.modes))
+    def unique_hash(h: primitive_harness.Harness, l: Jax2TfLimitation):
+      return (h.group_name, l.description, l.devices,
+              tuple([np.dtype(d).name for d in l.dtypes]), l.modes)
 
-    unique_limitations: Dict[Any, Jax2TfLimitation] = {}
+    unique_limitations: Dict[Any, Tuple[primitive_harness.Harness, Jax2TfLimitation]] = {}
     for h in harnesses:
       for l in Jax2TfLimitation.limitations_for_harness(h):
-        if l.expect_tf_error:
-          # We do not filter the limitations, we want them all
-          unique_limitations[unique_hash(l)] = l
+        unique_limitations[hash(unique_hash(h, l))] = (h, l)
 
     print(f"Found {len(unique_limitations)} unique limitations")
-    limited_support_table = [
+    tf_error_table = [
         """
-  | Affected primitive | Description of limitation | Affected dtypes | Affected devices | Affected compilation modes |
-  | --- | --- | --- | --- | --- | ---|"""
+| Affected primitive | Description of limitation | Affected dtypes | Affected devices | Affected compilation modes |
+| --- | --- | --- | --- | --- | ---|"""
     ]
-    for l in sorted(
-        unique_limitations.values(), key=lambda l: str(l.harness.group_name)):
+    tf_numerical_discrepancies_table = list(tf_error_table)  # a copy
+    for h, l in sorted(
+        unique_limitations.values(), key=lambda pair: unique_hash(*pair)):
       devices = ", ".join(sorted(l.devices))
       modes = ", ".join(sorted(l.modes))
       description = l.description
@@ -178,8 +176,16 @@ class JaxPrimitiveTest(tf_test_util.JaxToTfTestCase):
       if l.skip_tf_run:
         description = "TF test skipped: " + description
 
-      limited_support_table.append(
-          f"| {l.harness.group_name} | {description} | {primitive_harness.dtypes_to_str(l.dtypes, empty_means_all=True)} | {devices} | {modes} |"
+      if l.skip_tf_run or l.expect_tf_error:
+        to_table = tf_error_table
+      elif l.skip_comparison or l.custom_assert:
+        to_table = tf_numerical_discrepancies_table
+      else:
+        continue
+
+      to_table.append(
+          f"| {h.group_name} | {description} | "
+          f"{primitive_harness.dtypes_to_str(l.dtypes, empty_means_all=True)} | {devices} | {modes} |"
       )
 
     if not os.environ.get("JAX_OUTPUT_LIMITATIONS_DOC"):
@@ -203,7 +209,9 @@ class JaxPrimitiveTest(tf_test_util.JaxToTfTestCase):
 
     with open(output_file, "w") as f:
       f.write(template.replace("{{generation_date}}", str(datetime.date.today())) \
-              .replace("{{limited-support-table}}", "\n".join(limited_support_table)))
+              .replace("{{tf_error_table}}", "\n".join(tf_error_table)) \
+              .replace("{{tf_numerical_discrepancies_table}}", "\n".join(tf_numerical_discrepancies_table)) \
+              )
 
   # The rest of the test are checking special cases
 
@@ -304,1401 +312,6 @@ class JaxPrimitiveTest(tf_test_util.JaxToTfTestCase):
         tf_fn_array(np.array([3, 4, 5])), np.array([4.5, 10, 17.5],
                                                    jnp.bfloat16))
 
-
-class Jax2TfLimitation(primitive_harness.Limitation):
-  """Specific primitive limitations for jax2tf.
-
-  See the module docstring for details.
-  """
-
-  def __init__(
-      self,
-      harness: primitive_harness.Harness,
-      description: str,
-      *,
-      devices: Sequence[str] = ("cpu", "gpu", "tpu"),
-      dtypes=(),
-      enabled: bool = True,
-      # jax2tf specific
-      modes=("eager", "graph", "compiled"),
-      expect_tf_error: bool = True,
-      skip_tf_run=False,
-      custom_assert: Optional[Callable] = None,
-      skip_comparison=False,
-      tol=None):
-    """See the primitive_harness.Limitation common arguments.
-
-    Args :
-      modes: one of "eager", "graph", "compiled"
-      expect_tf_error: if set, then expect a TF error in the given mode when
-        executing the result of jax2tf conversion. If not set, then the
-        limitation must have a custom_assert or non-default tol.
-      skip_tf_run: if set will skip the TF execution. Use this sparingly,
-        prefer `expect_tf_error`. Use only when the test cannot recover from
-        the TF error.
-      custom_assert: if given, then execute as
-        `custom_assert(tst, result_jax, result_tf, args=args, tol=tol)`, where
-        `tst` is
-        the current TestCase instance, and args are the input arguments that
-        the harness created. The `tol` is the maximum tolerance based on the
-        limitations encountered so far.
-        `result_tf` is already converted to NumPy arrays.
-      tol: a tolerance to use for both atol and rtol.
-      skip_comparison: skips the numeric comparison. (TODO: implement this
-        with a custom assert)
-
-    """
-    assert isinstance(harness, primitive_harness.Harness)
-    super().__init__(
-        description,
-        devices=devices,
-        dtypes=dtypes,
-        harness=harness,
-        enabled=enabled)
-    if isinstance(modes, str):
-      modes = (modes,)
-    assert all(m in ["eager", "graph", "compiled"] for m in modes)
-    if devices == ("tpu",):
-      assert modes == ("compiled",), f"{modes}"
-    self.modes = modes
-    self.expect_tf_error = expect_tf_error
-    self.skip_tf_run = skip_tf_run
-    self.custom_assert = custom_assert
-    self.tol = tol
-    self.skip_comparison = skip_comparison
-
-  @classmethod
-  def limitations_for_harness(
-      cls, harness: primitive_harness.Harness) -> Sequence["Jax2TfLimitation"]:
-    group_method = getattr(cls, harness.group_name, None)
-    if harness.group_name in cls.harness_groups_no_limitations:
-      assert group_method is None, (
-          f"Harness group {harness.group_name} is both in "
-          f"'harness_groups_no_limitations' and has a custom "
-          f"Jax2TfLimitation.classmethod defined (see module docstring)"
-      )
-      return []
-    else:
-      assert group_method is not None, (
-          f"Harness group {harness.group_name} must be either part of "
-          f"'harness_groups_no_limitations' or must have a custom "
-          f"Jax2TfLimitation.classmethod defined (see module docstring)"
-      )
-      limitations = group_method(harness)
-      assert isinstance(limitations, (list, tuple))
-      return limitations
-
-  # We keep here the explicit set of groups for which we don't have limitations
-  harness_groups_no_limitations = {
-      "abs", "and", "argmin", "argmax", "broadcast", "broadcast_in_dim", "ceil",
-      "concatenate", "cos", "complex", "conj", "device_put", "dynamic_slice",
-      "dynamic_update_slice", "exp", "eq", "floor", "log", "gather", "imag",
-      "iota", "is_finite", "ne", "not", "or", "pad", "random_split",
-      "reduce_and", "reduce_prod", "reduce_or", "reduce_sum", "real", "reshape",
-      "select", "shift_left", "shift_right_logical", "shift_right_arithmetic",
-      "sin", "slice", "sqrt", "squeeze", "stop_gradient", "tie_in", "transpose",
-      "xor", "zeros_like"
-  }
-
-  @classmethod
-  def helper_get_trig_custom_limitation(cls, harness, np_inverse):
-
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      operand, = args
-      tst.assertAllClose(operand, np_inverse(result_tf), atol=tol, rtol=tol)
-
-    return custom_numeric(
-        harness,
-        dtypes=[np.complex64, np.complex128],
-        custom_assert=custom_assert,
-        also_compiled=True)
-
-  @classmethod
-  def acos(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16, np.complex64],
-            devices=("cpu", "gpu")),
-        missing_tf_kernel(
-            harness, dtypes=[np.complex128], devices=("cpu", "gpu")),
-        custom_numeric(harness, dtypes=np.complex128, tol=1e-13),
-        custom_numeric(harness, dtypes=np.complex64, devices="tpu", tol=1e-3),
-        custom_numeric(
-            harness, dtypes=np.complex64, devices=("cpu", "gpu"), tol=1e-4),
-        cls.helper_get_trig_custom_limitation(harness, np.cos),
-    ]
-
-  @classmethod
-  def acosh(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[dtypes.bfloat16, np.float16],
-            devices=("cpu", "gpu")),
-        custom_numeric(
-            harness, dtypes=np.complex64, devices=("cpu", "gpu"), tol=1e-3),
-        custom_numeric(
-            harness, dtypes=np.complex128, devices=("cpu", "gpu"), tol=1e-12),
-        cls.helper_get_trig_custom_limitation(harness, np.cosh)
-    ]
-
-  @classmethod
-  def add(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.uint16, np.uint32], also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64],
-            devices=("cpu", "gpu"),
-            also_compiled=True)
-    ]
-
-  @classmethod
-  # Also called add_jaxvals
-  def add_any(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint16, np.uint32, np.uint64],
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def asin(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16],
-            devices=("cpu", "gpu")),
-        missing_tf_kernel(
-            harness, dtypes=[np.complex64, np.complex128], also_compiled=True),
-        cls.helper_get_trig_custom_limitation(harness, np.sin)
-    ]
-
-  @classmethod
-  def asinh(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16],
-            devices=("cpu", "gpu")),
-        missing_tf_kernel(
-            harness, dtypes=[np.complex64, np.complex128], also_compiled=True),
-        custom_numeric(
-            harness, dtypes=np.complex64, devices=("cpu", "gpu"), tol=1e-3),
-        custom_numeric(
-            harness, dtypes=np.complex128, devices=("cpu", "gpu"), tol=1e-12),
-        cls.helper_get_trig_custom_limitation(harness, np.sinh)
-    ]
-
-  @classmethod
-  def atan(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16],
-            devices=("cpu", "gpu")),
-        missing_tf_kernel(
-            harness, dtypes=[np.complex64, np.complex128], also_compiled=True),
-        cls.helper_get_trig_custom_limitation(harness, np.tan)
-    ]
-
-  @classmethod
-  def atanh(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16],
-            devices=("cpu", "gpu")),
-        custom_numeric(harness, dtypes=np.float64, tol=1e-14),
-        custom_numeric(harness, dtypes=np.complex64, tol=1e-3),
-        custom_numeric(
-            harness, dtypes=np.complex128, devices=("cpu", "gpu"), tol=1e-12),
-        cls.helper_get_trig_custom_limitation(harness, np.tanh)
-    ]
-
-  @classmethod
-  def atan2(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16],
-            devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def bessel_i0e(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def bessel_i1e(cls, harness: primitive_harness.Harness):
-    return cls.bessel_i0e(harness)
-
-  @classmethod
-  def bitcast_convert_type(cls, harness: primitive_harness.Harness):
-    return [missing_tf_kernel(harness, dtypes=[np.bool_], also_compiled=True)]
-
-  @classmethod
-  def cholesky(cls, harness: primitive_harness.Harness):
-
-    def custom_assert(tst, result_jax, result_tf, *, tol, **_):
-      # cholesky_p returns garbage in the strictly upper triangular part of the
-      # result, so we can safely ignore that part.
-      tst.assertAllClose(jnp.tril(result_jax), result_tf, atol=tol)
-
-    return [
-        # See https://github.com/google/jax/pull/3775#issuecomment-659407824;
-        Jax2TfLimitation(
-            harness,
-            "function not compilable",
-            dtypes=[np.complex64, np.complex128],
-            devices=("cpu", "gpu"),
-            modes="compiled"),
-        missing_tf_kernel(
-            harness,
-            # Interesting: on TPU, complex64 works in eager
-            # mode, but fails otherwise.
-            dtypes=[np.complex64, np.complex128],
-            devices="tpu",
-            modes="compiled"),
-        # TODO(bchetioui): very high discrepancy in the float32/complex64 case
-        custom_numeric(harness, dtypes=[np.float32, np.complex64], tol=1e-2),
-        custom_numeric(harness, dtypes=[np.float64, np.complex128], tol=1e-6),
-        custom_numeric(harness, dtypes=[dtypes.bfloat16, np.float16], tol=5e-2),
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=True)
-    ]
-
-  @classmethod
-  def clamp(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.int8, np.uint16, np.uint32, np.uint64],
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def convert_element_type(cls, harness: primitive_harness.Harness):
-    return []
-
-  @classmethod
-  def conv_general_dilated(cls, harness: primitive_harness.Harness):
-    return [
-        # It is not enough to add an expect_tf_error limitation for this, because
-        # when this test runs with the LLVM debug-mode, it segfaults.
-        #   assert.h assertion failed at third_party/llvm/llvm-project/llvm/lib/IR/Instructions.cpp:2471
-        #      in void llvm::BinaryOperator::AssertOK(): getType()->isFPOrFPVectorTy() &&
-        #      "Tried to create a floating-point operation on a " "non-floating-point type!"
-        Jax2TfLimitation(
-            harness,
-            "TODO: TF/LLVM crash",
-            devices="gpu",
-            dtypes=[np.complex64, np.complex128],
-            skip_tf_run=True,
-        ),
-        Jax2TfLimitation(
-            harness,
-            "jax2tf BUG: batch_group_count > 1 not yet converted",
-            enabled=(harness.params["batch_group_count"] > 1)),
-        Jax2TfLimitation(
-            harness,
-            "XLA bug in the HLO -> LLVM IR lowering",
-            dtypes=[np.complex64, np.complex128],
-            devices=("cpu", "gpu"),
-            modes=("eager", "graph", "compiled")),
-        custom_numeric(harness, devices="gpu", tol=1e-4),
-        custom_numeric(harness, devices="tpu", tol=1e-3),
-        # TODO(bchetioui): significant discrepancies in some float16 cases.
-        custom_numeric(harness, dtypes=np.float16, tol=1),
-        # TODO(bchetioui): slight occasional discrepancy in float32 cases.
-        custom_numeric(harness, dtypes=np.float32, devices="tpu", tol=0.5),
-        custom_numeric(harness, dtypes=np.float32, devices="gpu", tol=1e-3),
-        custom_numeric(harness, dtypes=np.float32, devices="cpu", tol=1e-4),
-        custom_numeric(harness, dtypes=np.complex64, devices="tpu", tol=0.1),
-        # TODO(bchetioui): slight discrepancy when going through the path using
-        # tf.nn.convolution.
-        custom_numeric(harness, dtypes=np.float64, devices="cpu", tol=1e-13),
-    ]
-
-    # TODO(bchetioui): unidentified bug in compiled mode. The test that fails is
-    #
-    # test_conv_general_dilated_tf_conversion_path_3d_lhs=float32[1,4,28,28,1]_rhs=float32[2,3,3,1,16]_windowstrides=(1,1,1)_padding=VALID_lhsdilation=(1,1,1)_rhsdilation=(1,1,2)_dimensionnumbers=('NDHWC','DHWIO','NDHWC')_featuregroupcount=1_batchgroupcount=1_precision=None_enablexla=False
-    #
-    # with the following assertion error in TensorFlowTrace.process_primitive:
-    #
-    # AssertionError: conv_general_dilated: out.aval = ShapedArray(float32[1,3,24,26,16]); expected ShapedArray(float32[1,3,26,24,16])
-    #
-    # Deactivating this assertion is enough to pass the test, which suggests
-    # that the end shape is indeed the correct one (i.e. (1,3,26,24,16)).
-    # Further investigation is required to really understand this behavior,
-    # which we have not managed to reproduce as a pure TF test.
-    #
-    # This bug is low priority since it only occurs when using a non-TFXLA
-    # conversion path in compiled mode, i.e. in a context where using the
-    # TFXLA path is possible.
-    # if harness.name == "_tf_conversion_path_3d_lhs=float32[1,4,28,28,1]_rhs=float32[2,3,3,1,16]_windowstrides=(1,1,1)_padding=VALID_lhsdilation=(1,1,1)_rhsdilation=(1,1,2)_dimensionnumbers=('NDHWC','DHWIO','NDHWC')_featuregroupcount=1_batchgroupcount=1_precision=None_enablexla=False":
-    #  raise unittest.SkipTest("TODO: known but unidentified bug in compiled "
-    #                          "mode")
-
-  @classmethod
-  def cosh(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(harness, dtypes=[np.float16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def cummax(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64, np.complex128],
-            devices=("cpu", "gpu"),
-            also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint16, np.uint32, np.int8, np.complex64],
-            also_compiled=True),
-        custom_numeric(harness, dtypes=np.float16, tol=0.1),
-        custom_numeric(harness, dtypes=dtypes.bfloat16, tol=0.5)
-    ]
-
-  @classmethod
-  def cummin(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64, np.complex128],
-            devices=("cpu", "gpu"),
-            also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint16, np.uint32, np.int8, np.complex64],
-            also_compiled=True),
-        custom_numeric(harness, dtypes=np.float16, tol=0.1),
-        custom_numeric(harness, dtypes=dtypes.bfloat16, tol=0.5),
-    ]
-
-  @classmethod
-  def cumprod(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64],
-            devices=("cpu", "gpu"),
-            also_compiled=True),
-        missing_tf_kernel(harness, dtypes=[np.uint32], also_compiled=True),
-        custom_numeric(harness, dtypes=np.float16, tol=0.1),
-        custom_numeric(harness, dtypes=dtypes.bfloat16, tol=0.5),
-    ]
-
-  @classmethod
-  def cumsum(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64],
-            devices=("cpu", "gpu"),
-            also_compiled=True),
-        missing_tf_kernel(
-            harness, dtypes=[np.complex64], devices="tpu", modes="compiled"),
-        missing_tf_kernel(
-            harness, dtypes=[np.uint16, np.uint32], also_compiled=True),
-        custom_numeric(harness, dtypes=np.float16, tol=0.1),
-        custom_numeric(harness, dtypes=dtypes.bfloat16, tol=0.5),
-    ]
-
-  @classmethod
-  def custom_linear_solve(cls, harness: primitive_harness.Harness):
-    return [
-        Jax2TfLimitation(
-            harness,
-            "TODO: large numerical discrepancy",
-            dtypes=np.float32,
-            devices="tpu",
-            modes="compiled",
-            expect_tf_error=False,
-            skip_comparison=True),
-        custom_numeric(harness, dtypes=np.float32, devices="tpu", tol=0.01),
-        custom_numeric(harness, tol=1e-3),
-    ]
-
-  @classmethod
-  def digamma(cls, harness: primitive_harness.Harness):
-    dtype = harness.dtype
-
-    # In the bfloat16 case, TF and lax both return NaN in undefined cases.
-    # digamma is not defined at 0 and -1
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      # lax.digamma returns NaN and tf.math.digamma returns inf
-      arg, = args
-      special_cases = (arg == 0.) | (arg == -1.)
-      nr_special_cases = np.count_nonzero(special_cases)
-      tst.assertAllClose(
-          np.full((nr_special_cases,), dtype(np.nan)),
-          result_jax[special_cases])
-      tst.assertAllClose(
-          np.full((nr_special_cases,), dtype(np.inf)), result_tf[special_cases])
-      # non-special cases are equal
-      tst.assertAllClose(
-          result_jax[~special_cases],
-          result_tf[~special_cases],
-          atol=tol,
-          rtol=tol)
-
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu")),
-        custom_numeric(harness, dtypes=np.float64, tol=1e-13),
-        custom_numeric(
-            harness, dtypes=np.float32, devices=["cpu", "gpu"], tol=1e-3),
-        custom_numeric(
-            harness, dtypes=dtypes.bfloat16, custom_assert=custom_assert)
-    ]
-
-  @classmethod
-  def div(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.uint8, np.uint16, np.uint32, np.uint64, np.int8, np.int16
-            ],
-            also_compiled=True),
-        Jax2TfLimitation(
-            harness,
-            "TF integer division fails if divisor contains 0; JAX returns NaN",
-            dtypes=[
-                np.uint8, np.int8, np.uint16, np.uint32, np.uint64, np.int8,
-                np.int16, np.int32, np.int64
-            ],
-            # Only the harnesses with "singularity" will have divide by 0
-            enabled=("singularity" in harness.name))
-    ]
-
-  @classmethod
-  def dot_general(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.bool_, np.uint8, np.uint16, np.uint32, np.uint64, np.int8,
-                np.int16
-            ],
-            also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.int64],
-            devices=("cpu", "gpu"),
-            modes="compiled"),
-        custom_numeric(harness, dtypes=dtypes.bfloat16, tol=0.3),
-        custom_numeric(
-            harness,
-            dtypes=[np.complex64, np.float32],
-            devices=("cpu", "gpu"),
-            tol=1e-5),
-        custom_numeric(harness, dtypes=np.float32, devices="tpu", tol=0.1),
-        custom_numeric(harness, dtypes=np.complex64, devices="tpu", tol=0.3),
-        custom_numeric(
-            harness, dtypes=np.float16, devices=("gpu", "tpu"), tol=0.1),
-        custom_numeric(harness, dtypes=np.float16, devices="cpu", tol=0.01)
-    ]
-
-  @classmethod
-  def eig(cls, harness: primitive_harness.Harness):
-    compute_left_eigenvectors = harness.params["compute_left_eigenvectors"]
-    compute_right_eigenvectors = harness.params["compute_right_eigenvectors"]
-    dtype = harness.dtype
-
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      operand, = args
-      inner_dimension = operand.shape[-1]
-
-      # Test ported from tests.linlag_test.testEig
-      # Norm, adjusted for dimension and type.
-      def norm(x):
-        norm = np.linalg.norm(x, axis=(-2, -1))
-        return norm / ((inner_dimension + 1) * jnp.finfo(dtype).eps)
-
-      def check_right_eigenvectors(a, w, vr):
-        tst.assertTrue(
-            np.all(norm(np.matmul(a, vr) - w[..., None, :] * vr) < 100))
-
-      def check_left_eigenvectors(a, w, vl):
-        rank = len(a.shape)
-        aH = jnp.conj(a.transpose(list(range(rank - 2)) + [rank - 1, rank - 2]))
-        wC = jnp.conj(w)
-        check_right_eigenvectors(aH, wC, vl)
-
-      def check_eigenvalue_is_in_array(eigenvalue, eigenvalues_array):
-        tol = None
-        # TODO(bchetioui): numerical discrepancies
-        if dtype in [np.float32, np.complex64]:
-          tol = 1e-4
-        elif dtype in [np.float64, np.complex128]:
-          tol = 1e-13
-        closest_diff = min(abs(eigenvalues_array - eigenvalue))
-        tst.assertAllClose(
-            closest_diff, np.array(0., closest_diff.dtype), atol=tol)
-
-      all_w_jax, all_w_tf = result_jax[0], result_tf[0]
-      for idx in itertools.product(*map(range, operand.shape[:-2])):
-        w_jax, w_tf = all_w_jax[idx], all_w_tf[idx]
-        for i in range(inner_dimension):
-          check_eigenvalue_is_in_array(w_jax[i], w_tf)
-          check_eigenvalue_is_in_array(w_tf[i], w_jax)
-
-      if compute_left_eigenvectors:
-        check_left_eigenvectors(operand, all_w_tf, result_tf[1])
-      if compute_right_eigenvectors:
-        check_right_eigenvectors(operand, all_w_tf,
-                                 result_tf[1 + compute_left_eigenvectors])
-
-    return [
-        Jax2TfLimitation(harness, "function not compilable", modes="compiled"),
-        Jax2TfLimitation(
-            harness,
-            "TF Conversion of eig is not implemented when both compute_left_eigenvectors and compute_right_eigenvectors are set to True",
-            enabled=(compute_left_eigenvectors and compute_right_eigenvectors)),
-        custom_numeric(harness, custom_assert=custom_assert)
-    ]
-
-  @classmethod
-  def eigh(cls, harness: primitive_harness.Harness):
-    dtype = harness.dtype
-    shape = harness.params["shape"]
-
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      operand, = args
-      inner_dimension = operand.shape[-1]
-
-      def check_right_eigenvectors(a, w, vr):
-        tol = 1e-16
-        # TODO(bchetioui): tolerance needs to be very high in compiled mode,
-        # specifically for eigenvectors.
-        if dtype == np.float64:
-          tol = 1e-6
-        elif dtype == np.float32:
-          tol = 1e-2
-        elif dtype in [dtypes.bfloat16, np.complex64]:
-          tol = 1e-3
-        elif dtype == np.complex128:
-          tol = 1e-13
-        tst.assertAllClose(
-            np.matmul(a, vr) - w[..., None, :] * vr,
-            np.zeros(a.shape, dtype=vr.dtype),
-            atol=tol)
-
-      def check_eigenvalue_is_in_array(eigenvalue, eigenvalues_array):
-        tol = None
-        if dtype in [dtypes.bfloat16, np.float32, np.complex64]:
-          tol = 1e-3
-        elif dtype in [np.float64, np.complex128]:
-          tol = 1e-11
-        closest_diff = min(abs(eigenvalues_array - eigenvalue))
-        tst.assertAllClose(
-            closest_diff, np.array(0., closest_diff.dtype), atol=tol)
-
-      _, all_w_jax = result_jax
-      all_vr_tf, all_w_tf = result_tf
-
-      for idx in itertools.product(*map(range, operand.shape[:-2])):
-        w_jax, w_tf = all_w_jax[idx], all_w_tf[idx]
-        for i in range(inner_dimension):
-          check_eigenvalue_is_in_array(w_jax[i], w_tf)
-          check_eigenvalue_is_in_array(w_tf[i], w_jax)
-
-      check_right_eigenvectors(operand, all_w_tf, all_vr_tf)
-
-    return [
-        # See https://github.com/google/jax/pull/3775#issuecomment-659407824;
-        Jax2TfLimitation(
-            harness,
-            "function not compilable",
-            dtypes=[np.complex64, np.complex128],
-            modes="compiled",
-            enabled=(shape[0] > 0)),
-        Jax2TfLimitation(
-            harness,
-            "TODO: numeric discrepancies",
-            dtypes=[np.float64],
-            modes="compiled",
-            devices=("cpu", "gpu"),
-            expect_tf_error=False,
-            skip_comparison=True),
-        Jax2TfLimitation(
-            harness,
-            "TODO: numeric discrepancies",
-            dtypes=[np.float16],
-            modes="compiled",
-            devices=("tpu",),
-            expect_tf_error=False,
-            skip_comparison=True),
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=True)
-    ]
-
-  @classmethod
-  def ge(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(harness, dtypes=[np.bool_], also_compiled=True),
-        missing_tf_kernel(
-            harness, dtypes=[np.uint16, np.uint32], devices=("cpu", "gpu")),
-        missing_tf_kernel(harness, dtypes=[np.uint64], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def gt(cls, harness: primitive_harness.Harness):
-    return cls.ge(harness)
-
-  @classmethod
-  def erf(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def erfc(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def erf_inv(cls, harness: primitive_harness.Harness):
-    # erf_inv is not defined for arg <= -1 or arg >= 1
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):  # noqa: F811
-      arg, = args
-      # for arg < -1 or arg > 1
-      # lax.erf_inv returns NaN; tf.math.erf_inv return +/- inf
-      special_cases = (arg < -1.) | (arg > 1.)
-      # non-special cases are equal
-      tst.assertAllClose(
-          result_jax[~special_cases],
-          result_tf[~special_cases],
-          atol=tol,
-          rtol=tol)
-
-    return [
-        Jax2TfLimitation(harness,
-                         "TODO: erf_inv bug on TPU: nan vs non-nan",
-                         dtypes=np.float32,
-                         devices="tpu",
-                         skip_tf_run=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[dtypes.bfloat16, np.float16],
-            devices=("cpu", "gpu")),
-        custom_numeric(harness, dtypes=[np.float32, np.float64], tol=1e-4),
-        custom_numeric(
-            harness,
-            dtypes=[np.float32, np.float64],
-            custom_assert=custom_assert,
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def expm1(cls, harness: primitive_harness.Harness):
-    return [custom_numeric(harness, dtypes=np.float64, tol=1e-5)]
-
-  @classmethod
-  def fft(cls, harness):
-    return [
-        Jax2TfLimitation(
-            harness,
-            "TF function not compileable",
-            devices=("cpu", "gpu"),
-            dtypes=[np.float64, np.complex128],
-            modes="compiled"),
-        custom_numeric(harness, tol=1e-3)
-    ]
-
-  @classmethod
-  def _pow_test_util(cls, harness: primitive_harness.Harness):
-
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      # NaNs are mismatched, but assertAllClose will also behave weirdly for
-      # complex numbers containing np.inf as one of their components. See
-      # https://github.com/numpy/numpy/issues/15959 for more details.
-      mask = (
-          np.isnan(result_jax) + np.isnan(result_tf) + np.isinf(result_jax) +
-          np.isinf(result_tf))
-      tst.assertAllClose(result_jax[~mask], result_tf[~mask], rtol=tol)
-
-    return [
-        custom_numeric(
-            harness, dtypes=[np.float32, np.complex64], devices="tpu",
-            tol=1e-2),
-        custom_numeric(
-            harness,
-            dtypes=[np.float32, np.complex64],
-            devices=("cpu", "gpu"),
-            tol=1e-3),
-        custom_numeric(harness, dtypes=[np.float64, np.complex128], tol=1e-12),
-        custom_numeric(harness, dtypes=np.float16, tol=1),
-        # Values get really small for large negative powers.
-        custom_numeric(harness, dtypes=dtypes.bfloat16, tol=3),
-        custom_numeric(
-            harness,
-            dtypes=[np.complex64, np.complex128],
-            custom_assert=custom_assert,
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def igamma(cls, harness: primitive_harness.Harness):
-    dtype = harness.dtype
-
-    # igamma is not defined when the first argument is <=0
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      arg1, arg2 = args
-      # lax.igamma returns NaN when arg1 == arg2 == 0; tf.math.igamma returns 0
-      special_cases = (arg1 == 0.) & (arg2 == 0.)
-      nr_special_cases = np.count_nonzero(special_cases)
-      tst.assertAllClose(
-          np.full((nr_special_cases,), np.nan, dtype=dtype),
-          result_jax[special_cases])
-      tst.assertAllClose(
-          np.full((nr_special_cases,), 0., dtype=dtype),
-          result_tf[special_cases])
-      # non-special cases are equal
-      tst.assertAllClose(result_jax[~special_cases], result_tf[~special_cases])
-
-    return [
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=False)
-    ]
-
-  @classmethod
-  def igammac(cls, harness: primitive_harness.Harness):
-    dtype = harness.dtype
-
-    # igammac is not defined when the first argument is <=0
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):  # noqa: F811
-      arg1, arg2 = args
-      # lax.igammac returns 1. when arg1 <= 0; tf.math.igammac returns NaN
-      special_cases = (arg1 <= 0.) | (arg2 <= 0)
-      nr_special_cases = np.count_nonzero(special_cases)
-      tst.assertAllClose(
-          np.full((nr_special_cases,), 1., dtype=dtype),
-          result_jax[special_cases])
-      tst.assertAllClose(
-          np.full((nr_special_cases,), np.nan, dtype=dtype),
-          result_tf[special_cases])
-      # non-special cases are equal
-      tst.assertAllClose(
-          result_jax[~special_cases],
-          result_tf[~special_cases],
-          atol=tol,
-          rtol=tol)
-
-    return [
-        # Jax2TfLimitation(
-        #     harness,
-        #     "TODO: nan vs. non-nan",
-        #     devices="tpu",
-        #     dtypes=np.float32,
-        #     modes="compiled",
-        #     expect_tf_error=False,
-        #     custom_assert=lambda *_, **__: True),
-        custom_numeric(harness, dtypes=np.float64, tol=1e-9),
-        custom_numeric(harness, devices="gpu", tol=1e-3),
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=False),
-    ]
-
-  @classmethod
-  def integer_pow(cls, harness: primitive_harness.Harness):
-    y = harness.params["y"]
-    return [
-               missing_tf_kernel(
-                   harness,
-                   dtypes=[
-                       np.uint8, np.uint16, np.int8, np.int16, np.uint32, np.uint64
-                   ],
-                   also_compiled=True),
-               # hitting rtol = nan
-               Jax2TfLimitation(
-                   harness,
-                   "TODO: large numeric difference",
-                   devices="tpu",
-                   dtypes=np.complex64,
-                   modes="compiled",
-                   enabled=(y in [1000, -1000]),
-                   expect_tf_error=False,
-                   skip_comparison=True),
-               Jax2TfLimitation(
-                   harness,
-                   "TODO: large numeric difference for overflow",
-                   dtypes=[np.int32, np.int64, np.float32],
-                   enabled=(y > 10),
-                   expect_tf_error=False,
-                   skip_comparison=True)
-           ] + list(cls._pow_test_util(harness))
-
-  @classmethod
-  def pow(cls, harness: primitive_harness.Harness):
-    return cls._pow_test_util(harness)
-
-  @classmethod
-  def le(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(harness, dtypes=[np.bool_], also_compiled=True),
-        missing_tf_kernel(
-            harness, dtypes=[np.uint16, np.uint32], devices=("cpu", "gpu")),
-        missing_tf_kernel(harness, dtypes=[np.uint64], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def lt(cls, harness: primitive_harness.Harness):
-    return cls.ge(harness)
-
-  @classmethod
-  def lgamma(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu")),
-        custom_numeric(harness, dtypes=np.float64, tol=1e-11),
-        custom_numeric(harness, dtypes=np.float32, tol=1e-3)
-    ]
-
-  @classmethod
-  def log1p(cls, harness: primitive_harness.Harness):
-    return [
-        custom_numeric(harness, dtypes=np.float64, tol=1e-10),
-        custom_numeric(harness, dtypes=np.float32, tol=1e-3)
-    ]
-
-  @classmethod
-  def lu(cls, harness: primitive_harness.Harness):
-    dtype = harness.dtype
-
-    def custom_assert(tst, result_jax, result_tf, *, args, tol):
-      operand, = args
-      lu, pivots, perm = result_tf
-      batch_dims = operand.shape[:-2]
-      m, n = operand.shape[-2], operand.shape[-1]
-
-      def _make_permutation_matrix(perm):
-        result = []
-        for idx in itertools.product(*map(range, operand.shape[:-1])):
-          result += [0 if c != perm[idx] else 1 for c in range(m)]
-        result = np.reshape(np.array(result, dtype=dtype), [*batch_dims, m, m])
-        return result
-
-      k = min(m, n)
-      l = jnp.tril(lu, -1)[..., :, :k] + jnp.eye(m, k, dtype=dtype)
-      u = jnp.triu(lu)[..., :k, :]
-      p_mat = _make_permutation_matrix(perm)
-
-      tst.assertArraysEqual(
-          lax.linalg.lu_pivots_to_permutation(pivots, m), perm)
-      tst.assertAllClose(
-          jnp.matmul(p_mat, operand), jnp.matmul(l, u), atol=tol, rtol=tol)
-
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.complex64], devices="tpu", modes="compiled"),
-        custom_numeric(
-            harness, dtypes=[np.float32, np.complex64], devices="tpu", tol=0.1),
-        custom_numeric(
-            harness,
-            dtypes=[np.float32, np.complex64],
-            devices=("cpu", "gpu"),
-            tol=1e-5),
-        custom_numeric(harness, dtypes=[np.float64, np.complex128], tol=1e-13),
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=True),
-    ]
-
-  @classmethod
-  def _min_max_test_util(cls, harness: primitive_harness.Harness):
-    # TODO(bchetioui): discrepancies between TF & JAX when comparing with NaN;
-    # JAX always returns NaN, while TF returns the value NaN is compared with.
-    def custom_assert(tst, result_jax, result_tf, **_):
-      mask = np.isnan(result_jax)
-      tst.assertAllClose(result_jax[~mask], result_tf[~mask])
-
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.bool_, np.int8, np.complex64,
-                    np.uint16, np.uint32, np.uint64],
-            also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.complex128],
-            devices=("cpu", "gpu"),
-            also_compiled=True),
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=True)
-    ]
-
-  @classmethod
-  def max(cls, harness: primitive_harness.Harness):
-    return cls._min_max_test_util(harness)
-
-  @classmethod
-  def min(cls, harness: primitive_harness.Harness):
-    return cls._min_max_test_util(harness)
-
-  @classmethod
-  def mul(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.uint32, np.uint64], also_compiled=True)
-    ]
-
-  @classmethod
-  def neg(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint8, np.uint16, np.uint32, np.uint64],
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def nextafter(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.float16, dtypes.bfloat16], also_compiled=True)
-    ]
-
-  @classmethod
-  def population_count(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.uint32, np.uint64], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def qr(cls, harness: primitive_harness.Harness):
-    # See https://github.com/google/jax/pull/3775#issuecomment-659407824;
-    #     # experimental_compile=True breaks for complex types.
-    # TODO: see https://github.com/google/jax/pull/3775#issuecomment-659407824.
-    # - for now, the performance of the HLO QR implementation called when
-    #   compiling with TF is expected to have worse performance than the
-    #   custom calls made in JAX.
-    return [
-        custom_numeric(harness, tol=1e-5),
-        missing_tf_kernel(
-            harness,
-            dtypes=[dtypes.bfloat16],
-            devices="tpu",
-            modes=("compiled",))
-    ]
-
-  @classmethod
-  def random_gamma(cls, harness: primitive_harness.Harness):
-    return [custom_numeric(harness, devices="tpu", tol=1e-3)]
-
-  @classmethod
-  def reduce_max(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(harness, dtypes=[np.complex64], also_compiled=True),
-        missing_tf_kernel(harness, dtypes=[np.complex128], also_compiled=True)
-    ]
-
-  @classmethod
-  def reduce_min(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(harness, dtypes=[np.complex64], also_compiled=True),
-        missing_tf_kernel(harness, dtypes=[np.complex128], also_compiled=True)
-    ]
-
-  @classmethod
-  def reduce_window_add(cls, harness):
-    assert "add" == harness.params["computation"].__name__
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.uint16, np.uint32], also_compiled=True),
-        missing_tf_kernel(
-            harness, dtypes=[np.complex64], devices="tpu", also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64],
-            devices=("cpu", "gpu"),
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def reduce_window_mul(cls, harness):
-    assert "mul" == harness.params["computation"].__name__
-    return [
-        missing_tf_kernel(harness, dtypes=[np.uint32], also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64],
-            devices=("cpu", "gpu"),
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def reduce_window_min(cls, harness):
-    assert "min" == harness.params["computation"].__name__
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint32, np.uint16, np.bool_, np.complex64, np.int8],
-            also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64, np.complex128],
-            devices=("cpu", "gpu"),
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def reduce_window_max(cls, harness):
-    assert "max" == harness.params["computation"].__name__
-    dtype = harness.dtype
-    init_value = harness.params["init_value"]
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint32, np.bool_, np.complex64],
-            also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64, np.complex128],
-            devices=("cpu", "gpu"),
-            also_compiled=True),
-        Jax2TfLimitation(
-            harness,
-            "TF kernel missing, except when the initial_value is the minimum for the dtype",
-            dtypes=[np.uint16, np.int8],
-            enabled=((dtype == np.uint16 and init_value != 0) or
-                     (dtype == np.int8 and init_value != -128)))
-    ]
-
-  @classmethod
-  def regularized_incomplete_beta(cls, harness: primitive_harness.Harness):
-    return [
-        custom_numeric(harness, dtypes=np.float64, tol=1e-14),
-        missing_tf_kernel(
-            harness, dtypes=[np.float16, dtypes.bfloat16], also_compiled=True)
-    ]
-
-  @classmethod
-  def rem(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.uint8, np.uint16, np.uint32, np.uint64, np.int8, np.int16
-            ],
-            also_compiled=True),
-        Jax2TfLimitation(
-            harness,
-            "TF integer division fails if divisor contains 0; JAX returns NaN",
-            dtypes=[
-                np.uint8, np.int8, np.uint16, np.uint32, np.uint64, np.int8,
-                np.int16, np.int32, np.int64
-            ],
-            # Only the harnesses with "singularity" will have divide by 0
-            enabled=("singularity" in harness.name)),
-        missing_tf_kernel(harness, dtypes=[np.float16])
-    ]
-
-  @classmethod
-  def rev(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.uint32, np.uint64], also_compiled=True)
-    ]
-
-  @classmethod
-  def round(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def rsqrt(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def scatter_add(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.int8, np.uint16, np.uint32, np.uint64, np.complex64, np.bool_
-            ],
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def scatter_max(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.int8, np.uint16, np.uint32, np.uint64, np.complex64,
-                np.complex128, np.bool_
-            ],
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def scatter_min(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.int8, np.uint16, np.uint32, np.complex64, np.bool_,
-                np.uint64, np.complex128
-            ],
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def scatter_mul(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.int8, np.uint16, np.uint32, np.uint64, np.complex64, np.bool_
-            ],
-            also_compiled=True)
-    ]
-
-  # lax.select_and_gather_add_p:
-  #   # TODO: the conversion is only supported for float16/float32 on CPU/GPU,
-  #   # and float16 on TPU. This is because we do not implement a precision
-  #   # reduction in the case where packing 2 n-bit values together results in
-  #   # more than the maximum number of bits allowed on the platform (64 on
-  #   # CPU/GPU, 32 on TPU). This could be fixed by implementing a variadic
-  #   # reduce_window in tfxla, or we can require the user to reduce the
-  #   # precision of their arrays manually based on the platform they run on.
-  #   devices_and_max_bits = [ (["CPU", "GPU"], 64)
-  #                          , (["TPU"], 32)
-  #                          ]
-  #   for devs, max_bits in devices_and_max_bits:
-  #     if dtypes.finfo(np_dtype).bits * 2 > max_bits:
-  #       # TODO: getting an exception "XLA encountered an HLO for which this
-  #       # rewriting is not implemented"
-  #       tf_unimpl(np_dtype, devs=devs)
-  @classmethod
-  def select_and_gather_add(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.float32], devices="tpu", modes="compiled"),
-        Jax2TfLimitation(
-            harness,
-            "jax2tf unimplemented",
-            dtypes=[np.float64],
-            devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def select_and_scatter_add(cls, harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[np.uint32, np.uint16], also_compiled=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64],
-            devices=("cpu", "gpu"),
-            also_compiled=True)
-    ]
-
-  @classmethod
-  def sign(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[
-                np.uint32, np.uint16, np.int16, np.int8, np.uint8, np.uint64
-            ],
-            modes=(
-                "eager",
-                "graph",
-                "compiled",
-            ))
-    ]
-
-  @classmethod
-  def sinh(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(harness, dtypes=[np.float16], devices=("cpu", "gpu"))
-    ]
-
-  @classmethod
-  def sort(cls, harness: primitive_harness.Harness):
-    return [
-        Jax2TfLimitation(
-            harness,
-            # I think that this is because TF is running on CPU even for GPU tests?
-            "TODO: TF non-stable multiple-array sort",
-            devices="gpu",
-            enabled=(len(harness.arg_descriptors) == 4 and not harness.params["is_stable"]),
-            expect_tf_error=False,
-            skip_comparison=True),
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.complex64, np.complex128],
-            devices=("cpu", "gpu")),
-        Jax2TfLimitation(
-            harness,
-            "TODO: XlaSort does not support more than 2 arrays",
-            enabled=harness.params["nb_arrays"] > 2),
-        Jax2TfLimitation(
-            harness,
-            "TODO: XlaSort does not support sorting axis",
-            enabled=harness.params["dimension"] !=
-                    len(np.shape(harness.arg_descriptors[0])) - 1)
-    ]
-
-  @classmethod
-  def sub(cls, harness):
-    return [missing_tf_kernel(harness, dtypes=[np.uint64], also_compiled=True)]
-
-  @classmethod
-  def svd(cls, harness: primitive_harness.Harness):
-    # TODO: slow test
-
-    def custom_assert(tst, r_jax, r_tf, *, args, tol):
-
-      def _reconstruct_operand(result, is_tf: bool):
-        # Reconstructing operand as documented in numpy.linalg.svd (see
-        # https://numpy.org/doc/stable/reference/generated/numpy.linalg.svd.html)
-        s, u, v = result
-        U = u[..., :s.shape[-1]]
-        V = v[..., :s.shape[-1], :]
-        S = s[..., None, :]
-        return jnp.matmul(U * S, V), s.shape, u.shape, v.shape
-
-      if harness.params["compute_uv"]:
-        r_jax_reconstructed = _reconstruct_operand(r_jax, False)
-        r_tf_reconstructed = _reconstruct_operand(r_tf, True)
-        tst.assertAllClose(
-            r_jax_reconstructed, r_tf_reconstructed, atol=tol, rtol=tol)
-      else:
-        tst.assertAllClose(r_jax, r_tf, atol=tol, rtol=tol)
-
-    return [
-        Jax2TfLimitation(
-            harness,
-            "function not compilable",
-            dtypes=[np.complex64, np.complex128],
-            devices=("cpu", "gpu"),
-            modes=("compiled",)),
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], devices="tpu", modes="compiled"),
-        custom_numeric(harness, tol=1e-4),
-        custom_numeric(
-            harness, custom_assert=custom_assert, also_compiled=True)
-    ]
-
-  @classmethod
-  def tan(cls, harness):
-    return [
-        custom_numeric(harness, dtypes=np.complex64, devices="tpu", tol=1e-4),
-        custom_numeric(
-            harness, dtypes=np.complex64, devices=("cpu", "gpu"), tol=1e-3),
-        custom_numeric(
-            harness, dtypes=np.complex128, devices=("cpu", "gpu"), tol=1e-12)]
-
-  @classmethod
-  def tanh(cls, harness):
-    return [
-        custom_numeric(harness, dtypes=np.complex128, tol=1e-7),
-        custom_numeric(harness, dtypes=np.complex64, tol=1e-4)]
-
-  @classmethod
-  def top_k(cls, harness):
-
-    def custom_assert(tst, result_jax, result_tf, **_):
-      assert len(result_jax) == len(result_tf)
-      # TODO: TF and JAX sort [inf, nan] differently.
-      first_arr_jax, first_arr_tf = result_jax[0], result_tf[0]
-      if np.all(first_arr_jax == first_arr_tf):
-        for arr_jax, arr_tf in zip(result_jax, result_tf):
-          tst.assertArraysEqual(arr_jax, arr_tf)
-      else:
-        mask_jax, mask_tf = np.isnan(first_arr_jax), np.isnan(first_arr_tf)
-        tst.assertArraysEqual(first_arr_jax[~mask_jax], first_arr_tf[~mask_tf])
-
-    return [
-        missing_tf_kernel(
-            harness,
-            dtypes=[np.uint64, np.int64],
-            devices=("cpu", "gpu"),
-            modes="compiled"),
-        custom_numeric(
-            harness,
-            dtypes=[np.float16, dtypes.bfloat16, np.float32, np.float64],
-            custom_assert=custom_assert,
-            also_compiled=True)]
-
-  @classmethod
-  def triangular_solve(cls, harness: primitive_harness.Harness):
-    return [
-        missing_tf_kernel(
-            harness, dtypes=[dtypes.bfloat16], also_compiled=True),
-        missing_tf_kernel(harness, dtypes=[np.float16]),
-        custom_numeric(harness, dtypes=np.float32, tol=5e-3)]
-
-
-def custom_numeric(
-    harness: primitive_harness.Harness,
-    *,
-    description="custom numeric comparison",
-    dtypes=(),  # All
-    modes=("eager", "graph"),
-    also_compiled=False,
-    devices=("cpu", "gpu", "tpu"),
-    custom_assert=None,
-    tol=None) -> Jax2TfLimitation:
-  if isinstance(modes, str):
-    modes = (modes,)
-  # TODO: remove this hack
-  if not custom_assert and tol is not None:
-    also_compiled = True
-  if also_compiled and "compiled" not in modes:
-    modes = modes + ("compiled",)
-
-  return Jax2TfLimitation(
-      harness,
-      description,
-      expect_tf_error=False,
-      dtypes=dtypes,
-      devices=devices,
-      modes=modes,
-      custom_assert=custom_assert,
-      tol=tol)
-
-
-def missing_tf_kernel(
-    harness: primitive_harness.Harness,
-    *,
-    dtypes,
-    modes=("eager", "graph"),
-    also_compiled=False,
-    devices=("cpu", "gpu", "tpu")) -> Jax2TfLimitation:
-  if isinstance(modes, str):
-    modes = (modes,)
-  if also_compiled and "compiled" not in modes:
-    modes = modes + ("compiled",)
-
-  return Jax2TfLimitation(
-      harness,
-      "op not defined for dtype",
-      dtypes=dtypes,
-      devices=devices,
-      modes=modes)
 
 
 if __name__ == "__main__":
